fix: secure storage unable to read data from keychain
In some cases, the app could be initialised before secure storage was ready to be read (`isProtectedDataAvailable` was false). This could lead to data seemingly being "lost", while in fact it was just the keychain not being ready to read the data yet.

This introduces a workaround that will wait for the keychain to be ready before fully initialising the app.